move parsing logic to separate `GrepWorker` class
